Refactor TimeService.
